Time: 0 ms (100%), Space: 7.8 MB (91.74%) - LeetHub

+++ 0062-unique-paths/0062-unique-paths.cpp
@@ -1,14 +1,14 @@
 class Solution {
 public:
-    int ncr(int n , int r){
-        int ans = 1;
-        for(int i = 1; i<=min(n-r,r); i++){
+    long long ncr(int n , int r){
+        long long ans = 1;
+        for(long long i = 1; i<=min(n-r,r); i++){
             ans = (ans * (n-i+1))/i;
         }
         return ans;
     }
     int uniquePaths(int m, int n) {
-        int tot = m+n-2;
+        long long tot = m+n-2;
 
         return ncr(tot,m-1);
     }
